Refactor to introduce commander
