+++ packages/project-client-plugin-php/package.json
@@ -27,6 +27,7 @@
   },
   "homepage": "https://github.com/at-ishikawa/project-client/project-client-plugin-node/README.md",
   "devDependencies": {
+    "babel-cli": "^6.24.1",
     "babel-core": "^6.24.1",
     "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.24.1",
