+++ packages/project-client/package.json
@@ -28,10 +28,12 @@
   },
   "homepage": "https://github.com/at-ishikawa/project-client#readme",
   "dependencies": {
+    "commander": "^2.11.0",
     "jsonfile": "^2.4.0",
     "rmdir": "^1.2.0"
   },
   "devDependencies": {
+    "babel-cli": "^6.24.1",
     "babel-core": "^6.24.1",
     "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.24.1",
